+++ packages/atomic-angular/projects/atomic-angular/src/lib/stencil-generated/atomic-angular.module.ts
@@ -1,10 +1,11 @@
 /* tslint:disable */
 /* auto-generated angular module */
 import {CommonModule} from '@angular/common';
-import {ModuleWithProviders, NgModule} from '@angular/core';
+import {APP_INITIALIZER, ModuleWithProviders, NgModule, Provider} from '@angular/core';
 
         
 import {
+AtomicAriaLive,
 AtomicBreadbox,
 AtomicCategoryFacet,
 AtomicColorFacet,
@@ -19,6 +20,7 @@ AtomicFormatNumber,
 AtomicFormatUnit,
 AtomicFrequentlyBoughtTogether,
 AtomicIcon,
+AtomicLayoutSection,
 AtomicLoadMoreResults,
 AtomicNoResults,
 AtomicNumericFacet,
@@ -58,6 +60,7 @@ AtomicSearchBox,
 AtomicSearchBoxQuerySuggestions,
 AtomicSearchBoxRecentQueries,
 AtomicSearchInterface,
+AtomicSearchLayout,
 AtomicSortDropdown,
 AtomicSortExpression,
 AtomicTableElement,
@@ -72,6 +75,7 @@ defineCustomElements(window);
 
         
 const DECLARATIONS = [
+AtomicAriaLive,
 AtomicBreadbox,
 AtomicCategoryFacet,
 AtomicColorFacet,
@@ -86,6 +90,7 @@ AtomicFormatNumber,
 AtomicFormatUnit,
 AtomicFrequentlyBoughtTogether,
 AtomicIcon,
+AtomicLayoutSection,
 AtomicLoadMoreResults,
 AtomicNoResults,
 AtomicNumericFacet,
@@ -125,6 +130,7 @@ AtomicSearchBox,
 AtomicSearchBoxQuerySuggestions,
 AtomicSearchBoxRecentQueries,
 AtomicSearchInterface,
+AtomicSearchLayout,
 AtomicSortDropdown,
 AtomicSortExpression,
 AtomicTableElement,
@@ -134,10 +140,38 @@ AtomicTimeframeFacet
 ]
 
         
+const shimTemplates = ()=> {
+  // Angular's renderer will add children to a <template> instead of to its
+  // content. This shim will force any children added to a <template> to be
+  // added to its content instead.
+  // https://github.com/angular/angular/issues/15557
+  const nativeAppend = HTMLTemplateElement && HTMLTemplateElement.prototype && HTMLTemplateElement.prototype.appendChild;
+  if(!nativeAppend) {
+    return;
+  }
+  HTMLTemplateElement.prototype.appendChild = function<T extends Node>(
+    childNode: T
+  ) {
+    if (this.content) {
+      return this.content.appendChild(childNode);
+    } else {
+      return <T>nativeAppend.apply(this, [childNode]);
+    }
+  };
+}
+
+        
+const SHIM_TEMPLATES_PROVIDER: Provider = {
+  provide: APP_INITIALIZER,
+  multi: true,
+  useValue: shimTemplates
+}
+
+        
 @NgModule({
   declarations: DECLARATIONS,
   exports: DECLARATIONS,
-  providers: [],
+  providers: [SHIM_TEMPLATES_PROVIDER],
   imports: [CommonModule],
 })
 export class AtomicAngularModule {