+++ packages/atomic/cypress/integration/no-results-selectors.ts
@@ -1,5 +1,0 @@
-import {AriaLiveSelectors} from './aria-live-selectors';
-
-export const NoResultsSelectors = {
-  liveRegion: () => AriaLiveSelectors.region('no-results'),
-};