+++ packages/atomic-react/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@0.5.0...@coveo/atomic-react@0.5.1) (2022-02-16)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+# [0.5.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@0.4.14...@coveo/atomic-react@0.5.0) (2022-02-15)
+
+
+### Features
+
+* **atomic:** add search layout ([#1730](https://github.com/coveo/ui-kit/issues/1730)) ([aed0cae](https://github.com/coveo/ui-kit/commit/aed0caea993016ed90b941c377ccd52895146246))
+
+
+
+
+
+## [0.4.14](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@0.4.13...@coveo/atomic-react@0.4.14) (2022-02-11)
+
+
+### Bug Fixes
+
+* **atomic:** improved aria-live accessibility ([#1715](https://github.com/coveo/ui-kit/issues/1715)) ([124531f](https://github.com/coveo/ui-kit/commit/124531f9325e81b1e0856001f4ec4b1bc2ce8664))
+
+
+
+
+
+## [0.4.13](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@0.4.12...@coveo/atomic-react@0.4.13) (2022-02-09)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [0.4.12](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@0.4.11...@coveo/atomic-react@0.4.12) (2022-02-07)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
 ## [0.4.11](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@0.4.10...@coveo/atomic-react@0.4.11) (2022-02-04)
 
 **Note:** Version bump only for package @coveo/atomic-react